requirements.txt
torch
transformers
opencv-python
numpy
pillow
tk
